// lib/pdf/templates/modern.d.ts
import React from "react";
import { CVData, ExportOptions } from "../types";
export interface ModernTemplateProps {
    data: CVData;
    options: ExportOptions;
}
export declare const ModernTemplate: ({ data, options }: ModernTemplateProps) => React.JSX.Element;
export default ModernTemplate;
//# sourceMappingURL=modern.d.ts.map